requirements.txt
amqp==5.3.1
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.7.1
cryptography==44.0.2
defusedxml==0.7.1
dj-database-url==2.3.0
Django==5.1.7
django-appconf==1.1.0
django-autoslug==1.9.9
django-celery-email==3.0.0
django-cors-headers==4.7.0
django-countries==7.6.1
django-environ==0.12.0
django-filter==25.1
django-phonenumber-field==8.0.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-yasg==1.21.10
factory_boy==3.3.3
Faker==37.1.0
flake8==7.1.2
flower==2.0.1
gunicorn==23.0.0
humanize==4.12.2
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
isort==6.0.1
kombu==5.5.1
mccabe==0.7.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.2
pathspec==0.12.1
phonenumbers==9.0.2
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycountry==24.6.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.5
pytest-cov==6.0.0
pytest-django==4.10.0
pytest-factoryboy==2.7.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tornado==6.4.2
typing_extensions==4.13.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
whitenoise==6.9.0
